chore: add php-cs-fixer and phpstan (#2)

+++ src/Service/FileSizeFormatter.php
@@ -10,7 +10,7 @@ class FileSizeFormatter
             return 'N/A';
         }
 
-        $i = floor(log($bytes) / log(1024));
+        $i = (int) floor(log($bytes) / log(1024));
         $units = ['B', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
 
         return sprintf('%.02F', $bytes / pow(1024, $i)) * 1 . ' ' . $units[$i];
